# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Experts api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/clients/trainings": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Returns list of client trainings",
                "description": "Returns list of client trainings",
                "operationId": "getAllClientTrainings",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succefully found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "limit": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "records": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "trainer_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "start_at": {
                                                        "type": "datetime",
                                                        "example": "2021-10-23 13:00:00"
                                                    },
                                                    "finish_at": {
                                                        "type": "datetime",
                                                        "example": "2021-10-23 15:00:00"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "created_at": {
                                                        "type": "datetime",
                                                        "example": "2021-10-08T15:32:51.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "datetime",
                                                        "example": "2021-10-08T18:43:44.000000Z"
                                                    },
                                                    "client_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "training_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/v1/trainings": {
            "get": {
                "tags": [
                    "Training"
                ],
                "summary": "Returns list of trainings",
                "description": "Returns list of trainings",
                "operationId": "getAllTrainings",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succefully found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "limit": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "trainers": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Training"
                ],
                "summary": "Trainer add new training",
                "description": "Trainer add new training",
                "operationId": "trainingAdd",
                "requestBody": {
                    "description": "Pass new customer data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "start_at": {
                                        "description": "Date format: Y-m-d H:i:s",
                                        "type": "string",
                                        "example": "2021-10-23 13:00:00"
                                    },
                                    "finish_at": {
                                        "description": "Date format: Y-m-d H:i:s",
                                        "type": "string",
                                        "example": "2021-10-23 15:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "training": {
                                            "properties": {
                                                "start_at": {
                                                    "type": "datetime",
                                                    "example": "2021-10-23 13:00:00"
                                                },
                                                "finish_at": {
                                                    "type": "datetime",
                                                    "example": "2021-10-23 15:00:00"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "trainer_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "updated_at": {
                                                    "type": "datetime",
                                                    "example": "2021-10-08T19:26:06"
                                                },
                                                "created_at": {
                                                    "type": "datetime",
                                                    "example": "2021-10-08T19:26:06"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "first_name",
                    "last_name",
                    "email",
                    "phone_number",
                    "password",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "example": "user@gmail.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "0990619649"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "trainer",
                            "client"
                        ],
                        "example": "trainer"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Harry"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Potter"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}